refactor: 컨벤션에 따른 네이밍 수정
