--- app/src/main/java/com/teamwss/websoso/ui/novel/novelDetail/NovelDetailFragment.kt
@@ -1,8 +1,0 @@
-package com.teamwss.websoso.ui.novel.novelDetail
-
-import com.teamwss.websoso.R
-import com.teamwss.websoso.databinding.FragmentNovelDetailBinding
-import com.teamwss.websoso.ui.common.base.BindingFragment
-
-class NovelDetailFragment :
-    BindingFragment<FragmentNovelDetailBinding>(R.layout.fragment_novel_detail) {}
