+++ app/src/main/java/com/teamwss/websoso/ui/novelDetail/NovelDetailActivity.kt
@@ -1,11 +1,11 @@
-package com.teamwss.websoso.ui.novel
+package com.teamwss.websoso.ui.novelDetail
 
 import android.os.Bundle
 import com.teamwss.websoso.R
-import com.teamwss.websoso.databinding.ActivityNovelBinding
+import com.teamwss.websoso.databinding.ActivityNovelDetailBinding
 import com.teamwss.websoso.ui.common.base.BindingActivity
 
-class NovelActivity: BindingActivity<ActivityNovelBinding>(R.layout.activity_novel) {
+class NovelDetailActivity: BindingActivity<ActivityNovelDetailBinding>(R.layout.activity_novel_detail) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
