+++ app/src/main/java/com/teamwss/websoso/ui/novelDetail/novelFeed/NovelFeedFragment.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.novel.novelFeed
+package com.teamwss.websoso.ui.novelDetail.novelFeed
 
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.FragmentNovelFeedBinding
