+++ app/src/main/java/com/teamwss/websoso/ui/novelDetail/novelInfo/NovelInfoFragment.kt
@@ -1,0 +1,8 @@
+package com.teamwss.websoso.ui.novelDetail.novelInfo
+
+import com.teamwss.websoso.R
+import com.teamwss.websoso.databinding.FragmentNovelInfoBinding
+import com.teamwss.websoso.ui.common.base.BindingFragment
+
+class NovelInfoFragment :
+    BindingFragment<FragmentNovelInfoBinding>(R.layout.fragment_novel_info) {}
