+++ app/src/main/res/layout/fragment_novel_info.xml
